+++ ile_de_france/src/main/java/org/eqasim/ile_de_france/policies/PolicyFactory.java
@@ -1,0 +1,5 @@
+package org.eqasim.ile_de_france.policies;
+
+public interface PolicyFactory {
+	Policy createPolicy(String name, PolicyPersonFilter personFilter);
+}